requirements.txt
transformers
python3-timbl
pandas
numpy
torch
codecarbon
flask
